requirements.txt
# 核心依赖
transformers>=4.35.0
peft>=0.7.0
datasets>=2.15.0
bitsandbytes>=0.41.0
accelerate>=0.25.0
torch>=2.0.0

# 分词和文本处理
sentencepiece>=0.1.99
jieba>=0.42.1

# 评估指标
rouge-chinese>=1.0.3

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 工具库
tqdm>=4.66.0
tensorboard>=2.15.0
pyyaml>=6.0.0
